Update themes and restructure assets folder
