+++ src/__configuration__/themes/schedule.ts
@@ -1,5 +1,5 @@
 import { HalloweenSchedule } from './halloween';
-import { MayTheForthSchedule } from './may-the-forth';
+import { MayTheFourthSchedule } from './may-the-fourth';
 import { ChristmasSchedule } from './christmas';
 import { defaultTheme } from './default';
 import { Schedule, SiteConfig } from './types';
@@ -15,13 +15,13 @@ import { NewYearsSchedule } from './new-years';
  *
  * Current schedules:
  * * Mar 7 - Mar 9: Women's Day
- * * May 3 - May 6: May the Forth Day
+ * * May 3 - May 6: May the Fourth Day
  * * Oct 14 - Nov 3: Halloween
  * * Nov 12 - Nov 17: Diwali (changes each year)
  * * Nov 22 - Nov 30: Thanksgiving (changes each year)
  * * Dec 10 - Dec 18: Hanukkah (changes each year)
  * * Dec 23 - Dec 26: Christmas
- * * Dec 26 - Dec 30: Kwanzaa (Kwanzaa runs until Jan 1. Cutting This short to account for New Years.)
+ * * Dec 26 - Dec 30: Kwanzaa (Kwanzaa runs until Jan 1. Cutting this short to account for New Years.)
  * * Dec 31 - Jan 1: New Years
  */
 export const schedule: Schedule[] = [
@@ -30,7 +30,7 @@ export const schedule: Schedule[] = [
     DiwaliSchedule,
     ChristmasSchedule,
     HalloweenSchedule,
-    MayTheForthSchedule,
+    MayTheFourthSchedule,
     HanukkahSchedule,
     KwanzaaSchedule,
     NewYearsSchedule,
