+++ src/app/assets/themes/christmas/christmas-appbar-tile.png
