+++ src/app/assets/themes/christmas/christmas-eve.png
