+++ src/app/assets/themes/diwali/diwali-appbar-tile.png
