+++ src/app/assets/themes/diwali/diwali-cursor.png
