+++ src/app/assets/themes/halloween/witch.png
