+++ src/app/assets/themes/thanksgiving/turkey-cursor.png
